requirements.txt
pandas
numpy
python-dotenv
openai
pyyaml
